# lib/api-zod/src/generated/types/deploymentListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky-Hosting API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Deployment } from "./deployment";

export interface DeploymentListResponse {
  deployments: Deployment[];
  total: number;
  limit: number;
  offset: number;
}