# lib/api-zod/src/generated/types/envVarsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky-Hosting API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnvVarsResponse {
  deploymentId: string;
  keys: string[];
}